Move agent-count segment tests to their own file for the length budget
